% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{arrange_rows}
\alias{arrange_rows}
\title{arrange_rows}
\usage{
arrange_rows(matrix)
}
\arguments{
\item{matrix}{The matrix in which you would like to rearrange the rows}
}
\value{
The matrix with its rows rearranged
}
\description{
arranges rows of matrix in descending order starting with the leftmost position of each row.
}
\examples{
\dontrun{arrange_rows(rbind(c(0, 0, 1), c(1, 0, 0), c(0, 1, 1)))}

}
