% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{checkSubspace}
\alias{checkSubspace}
\title{Define the function to check subspace}
\usage{
checkSubspace(vectors)
}
\arguments{
\item{vectors}{a list of vectors}
}
\description{
Define the function to check subspace
}
