% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-d.R
\name{d}
\alias{d}
\title{Deviation from the Mean}
\usage{
d(x)
}
\arguments{
\item{x}{Numeric matrix, data frame, or vector.
Data matrix.}
}
\value{
A matrix.
}
\description{
Calculates the deviation from the mean.
}
\details{
The
\eqn{n \times 1}
vector of deviations from the mean
is given by

\deqn{
    \mathbf{d}
    =
    \mathbf{x} - \bar{x}
}

where
\eqn{\mathbf{x}}
is an
\eqn{n \times 1}
column vector
and
\eqn{\bar{x}}
is the mean of
\eqn{\mathbf{x}}.

The
\eqn{n \times k}
matrix of deviations from the mean
is given by

\deqn{
    \mathbf{D}
    =
    \mathbf{X} - \mathbf{1}_{n} \bar{\mathbf{x}}
}

where
\eqn{\mathbf{X}}
is an
\eqn{n \times 1}
matrix,
\eqn{\mathbf{1}_{n}}
is an
\eqn{n \times 1}
column vector of ones,
and
\eqn{\bar{\mathbf{x}}}
is the
\eqn{k \times 1}
mean vector
of
\eqn{\mathbf{X}}.
}
\examples{
x <- matrix(
  rnorm(n = 5 * 5, mean = 100, sd = 15),
  ncol = 5
)

d(x)

x <- rnorm(5)

d(x)
}
\seealso{
Other Scaling Functions: 
\code{\link{deltacapsq}()},
\code{\link{z}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Scaling Functions}
\keyword{linearAlgebra}
\keyword{scaling}
