% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-dcap.R
\name{dcap}
\alias{dcap}
\title{Duplication Matrix}
\usage{
dcap(x)
}
\arguments{
\item{x}{Integer.
Dimension of the symmetric matrix.}
}
\value{
A matrix.
}
\description{
Create a duplication matrix.
}
\details{
The duplication matrix
\eqn{
\mathbf{D}_{k}
}
is the
\eqn{
k^2 \times \frac{k \left( k + 1 \right)}{2}
}
matrix
for a given
\eqn{
k \times k
}
symmetric matrix
\eqn{
\mathbf{A}
}
where

\deqn{
   \mathbf{D}_{k} \mathrm{vech} \left( \mathbf{A} \right)
   =
   \mathrm{vec} \left( \mathbf{A} \right)
}

\eqn{
\mathrm{vec} \left( \cdot \right)
}
is the vectorization of a matrix,
and
\eqn{
\mathrm{vech} \left( \cdot \right)
}
is the half-vectorization of a matrix.
}
\examples{
dcap(3)
}
\references{
\href{https://en.wikipedia.org/wiki/Duplication_and_elimination_matrices#Duplication_matrix}{Wikipedia: Duplication matrix}

Magnus, J. R., & Neudecker, H. (1980).
The elimination matrix: Some lemmas and applications.
SIAM  Journal  on  Algebraic  Discrete  Methods,
1(4),
422--449.
\url{https://doi.org/10.1137/0601049}

Magnus, J. R., & Neudecker, H. (2019).
Matrix  differential  calculus with  applications  in  statistics  and  econometrics.
Wiley.
\url{https://doi.org/10.1002/9781119541219}
}
\seealso{
Other Linear Algebra Functions: 
\code{\link{diag_of_vech}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()},
\code{\link{test_positive_definite}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Linear Algebra Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
