% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{def}
\alias{def}
\title{Return definition of a term in linear algebra}
\usage{
def(term)
}
\arguments{
\item{term}{a term that's commonly used in linear algebra}
}
\value{
a sentence of definition corresponding to the term,
or "Term not found" if it's not included in our glossory.
}
\description{
This function returns the definition of the input term in linear algebra,
if the term doesn't exist in our glossory term, it will return "Term not found"
}
\examples{
\dontrun{
def("Codomain")
}
}
