% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-deltacapsq.R
\name{deltacapsq}
\alias{deltacapsq}
\title{Squared Mahalanobis Distance}
\usage{
deltacapsq(x, mu, sigmacap)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{mu}{Numeric vector.
Center
\eqn{\boldsymbol{\mu}}.}

\item{sigmacap}{Numeric matrix.
Covariance matrix
\eqn{\boldsymbol{\Sigma}}.}
}
\value{
A vector.
}
\description{
Calculates the squared Mahalanobis distance.
}
\details{
The squared Mahalanobis distance is given by
\deqn{
    \boldsymbol{\Delta}^{2}
    =
    \left( \mathbf{x} - \boldsymbol{\mu} \right)^{\prime}
    \boldsymbol{\Sigma}^{-1}
    \left( \mathbf{x} - \boldsymbol{\mu} \right)
}
}
\examples{
x <- matrix(
  data = stats::rnorm(n = 5 * 2),
  ncol = 2
)

deltacapsq(
  x,
  mu = c(0, 0),
  sigmacap = diag(2)
)

x <- rnorm(5)

deltacapsq(
  x,
  mu = 0,
  sigmacap = 1
)
}
\references{
\href{https://en.wikipedia.org/wiki/Mahalanobis_distance}{Wikipedia: Mahalanobis distance}
}
\seealso{
Other Scaling Functions: 
\code{\link{d}()},
\code{\link{z}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Scaling Functions}
\keyword{linearAlgebra}
\keyword{scaling}
