% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{det}
\alias{det}
\title{det}
\usage{
det(matrix)
}
\arguments{
\item{matrix}{The matrix you want the determinant of}
}
\value{
The determinant of the given matrix
}
\description{
computes the determinant of a matrix
}
\note{
Number of rows must equal number of columns
}
\examples{
\dontrun{det(matrix(1:49, ncol = 7))}
\dontrun{det(i_matrix(5))}

}
