% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIM.R
\name{dimension}
\alias{dimension}
\title{Dimension}
\usage{
dimension(...)
}
\arguments{
\item{...}{R objects}
}
\value{
dimension of any R object
}
\description{
Dimension function which works for not only matrix and data.frame
}
