% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-mcap_sym-dot.R
\name{.mcap_sym}
\alias{.mcap_sym}
\title{The Symmetric Pattern Matrix - Dot Function}
\usage{
.mcap_sym(d)
}
\arguments{
\item{d}{Numeric matrix.
Duplication matrix.}
}
\value{
A matrix.
}
\description{
Creates a symmetric pattern matrix.
}
\details{
\eqn{
    \mathbf{M}_{k}
    \left(
    s
    \right)
}
is the
\eqn{
    k \times k
}
symmetric pattern matrix with

\deqn{
    \left(
	   \mathbf{M}_{k}
	   \left(
	   s
	   \right)
	   \right)_{ij, gh}
    =
	   \begin{cases}
	       1
		   &
		   \text{if}
		   \quad
		   i = j = g = h
	       ,
		   \\
		   \frac{1}{2}
		   &
		   \text{if}
		   \quad
		   \left( i, j \right) = \left( g, h \right)
		   \text{or}
		   \left( i, j \right) = \left( h, g \right)
		   ,
		   \\
		   0
		   &
		   \text{otherwise}
		   .
    \end{cases}
}
}
\examples{
.mcap_sym(3)
}
\references{
Nel, D. G. (1985).
A matrix derivation of the asymptotic covariance matrix of sample correlation coefficients.
Linear Algebra and its Applications,
67, 137--145.
https://doi.org/10.1016/0024-3795(85)90191-0
}
\seealso{
Other Symmetric Functions: 
\code{\link{.dcap}()},
\code{\link{.mcap_cor}()},
\code{\link{.mcap_diag}()},
\code{\link{.pinv_of_dcap}()},
\code{\link{.sym_of_vechs}()},
\code{\link{.sym_of_vech}()},
\code{\link{dcap}()},
\code{\link{k_check}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{pinv_of_dcap}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{dot}
\keyword{linearAlgebra}
\keyword{symmetric}
