% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-sym_of_vech-dot.R
\name{.sym_of_vech}
\alias{.sym_of_vech}
\title{Symmetric matrix A from vech(A) - Dot Function}
\usage{
.sym_of_vech(x, k)
}
\arguments{
\item{x}{Vector.
Half-vectorization of a \code{k} by \code{k} matrix.}

\item{k}{Positive integer.
Dimension of the \code{k} by \code{k} matrix.}
}
\value{
A matrix.
}
\description{
Symmetric matrix from its half-vectorization.
}
\details{
Generates an
\eqn{k \times k}
symmetric matrix
from a
\eqn{\frac{1}{2}k(k + 1)}
vector.
}
\examples{
A <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  nrow = 3
)
vechA <- c(1.0, 0.5, 0.4, 1.0, 0.6, 1.0)

sym_of_vech(vechA)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)}{Wikipedia: Vectorization}

Magnus, J. R., & Neudecker, H. (2019).
Matrix  differential  calculus with  applications  in  statistics  and  econometrics.
Wiley.
\url{https://doi.org/10.1002/9781119541219}
}
\seealso{
Other Symmetric Functions: 
\code{\link{.dcap}()},
\code{\link{.mcap_cor}()},
\code{\link{.mcap_diag}()},
\code{\link{.mcap_sym}()},
\code{\link{.pinv_of_dcap}()},
\code{\link{.sym_of_vechs}()},
\code{\link{dcap}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{pinv_of_dcap}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{dot}
\keyword{linearAlgebra}
\keyword{symmetric}
