% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vec-dot.R
\name{.vec}
\alias{.vec}
\title{Vectorize - Dot Function}
\usage{
.vec(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
A vector.
}
\description{
Vectorize a matrix.
}
\details{
The vectorization of an
\eqn{m \times n}
matrix
\eqn{\mathbf{A}},
given by
\eqn{
    \mathrm{vec} \left( \mathbf{A} \right)
},
is the
\eqn{mn \times 1}
vector obtained by stacking the elements of
\eqn{\mathbf{A}}
column-wise.
}
\examples{
A <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  nrow = 3
)

.vec(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)}{Wikipedia: Vectorization}

Magnus, J. R., & Neudecker, H. (2019).
Matrix  differential  calculus with  applications  in  statistics  and  econometrics.
Wiley.
\url{https://doi.org/10.1002/9781119541219}
}
\seealso{
Other Vectorization Functions: 
\code{\link{.diag_of_vech}()},
\code{\link{.vechnames}()},
\code{\link{.vechsnames}()},
\code{\link{.vechs}()},
\code{\link{.vech}()},
\code{\link{diag_of_vech}()},
\code{\link{vech_check}()},
\code{\link{vechnames}()},
\code{\link{vechs_check}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{dot}
\keyword{linearAlgebra}
\keyword{vectorization}
