% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vec_mean-dot.R
\name{.vec_mean}
\alias{.vec_mean}
\title{Vector Mean - Dot Function}
\usage{
.vec_mean(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Calculates the mean of a vector.
}
\details{
The vector mean is given by

\deqn{
  \frac{1}{n}
  \left(
  \mathbf{1}_{n}^{\prime}
  \mathbf{x}
  \right)
}
}
\seealso{
Other Vectorization Functions: 
\code{\link{.diag_of_vech}()},
\code{\link{.vechnames}()},
\code{\link{.vechsnames}()},
\code{\link{.vechs}()},
\code{\link{.vech}()},
\code{\link{.vec}()},
\code{\link{diag_of_vech}()},
\code{\link{vec_mean}()},
\code{\link{vech_check}()},
\code{\link{vechnames}()},
\code{\link{vechs_check}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{dot}
\keyword{linearAlgebra}
\keyword{operation}
