% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vechs-dot.R
\name{.vechs}
\alias{.vechs}
\title{Strict Half-Vectorize - Dot Function}
\usage{
.vechs(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
A vector.
}
\description{
Apply a strict half-vectorization,
that is,
eliminate the diagonal and upper diagonal elements of a
\eqn{k \times k}
matrix.
}
\details{
The strict half-vectorization of a
\eqn{k \times k}
matrix
\eqn{\mathbf{A}},
given by
\eqn{
    \mathrm{vechs} \left( \mathbf{A} \right)
},
is the
\eqn{
    \frac{1}{2}
    k
    \left(
    k + 1
    \right)
    -
    k
    \times
    1
}
vector obtained from
the vectorization of
\eqn{\mathbf{A}},
given by
\eqn{
    \mathrm{vec} \left( \mathbf{A} \right)
},
where that all diagonal and upper diagonal elements of
\eqn{\mathbf{A}}
are eliminated.
}
\examples{
A <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  ncol = 3
)

.vechs(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)}{Wikipedia: Vectorization}

Magnus, J. R., & Neudecker, H. (2019).
Matrix  differential  calculus with  applications  in  statistics  and  econometrics.
Wiley.
\url{https://doi.org/10.1002/9781119541219}
}
\seealso{
Other Vectorization Functions: 
\code{\link{.diag_of_vech}()},
\code{\link{.vechnames}()},
\code{\link{.vechsnames}()},
\code{\link{.vech}()},
\code{\link{.vec}()},
\code{\link{diag_of_vech}()},
\code{\link{vech_check}()},
\code{\link{vechnames}()},
\code{\link{vechs_check}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{dot}
\keyword{linearAlgebra}
\keyword{vectorization}
