% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{findBasisFromList}
\alias{findBasisFromList}
\title{find the basis of a set of vectors}
\usage{
findBasisFromList(vectors)
}
\arguments{
\item{vectors}{a set of vectors}
}
\value{
This returns the set of vectors that's the basis
}
\description{
Given list of vectors, this identify the linearly independent vectors and
returns the basis of the vector.
}
\examples{
vectors <- list(
a = c(1, 2, 3),
b = c(4, 5, 6),
c = c(7, 8, 9)
)
findBasisFromList(vectors)
}
