% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixReduceProcess.R
\name{GaussianElimination}
\alias{GaussianElimination}
\title{Gaussian Elimination}
\usage{
GaussianElimination(coefMatrix, attachVector, FRAC = TRUE, PRINT = FALSE)
}
\arguments{
\item{coefMatrix}{ceofficient matrix}

\item{attachVector}{An additional column which attach to the coefficient matrix => augmented matrix}

\item{FRAC}{Flag of showing result by fraction (Default is TRUE)}

\item{PRINT}{Flag of printing process detail (Default is FALSE)}
}
\value{
Return a List with StrictlyTriangularForm; AttachVector
}
\description{
Gaussian Elimination: Reduce Augmented Matrix to Row Echelon Form
Definition: A matrix is said to be in row echelon form
(i)	If the first nonzero entry in each nonzero row is 1.
(ii)	If row k does not consist entirely of zeros, the number of leading zero entries in row k + 1 is greater than the number of leading zero entries in row k.
(iii)	If there are rows whose entries are all zero, they are below the row shaving nonzero entries.
Using Row Operations I, II, III
}
