% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glossary}
\alias{glossary}
\title{Glossory of linear algebra terms}
\format{
\describe{
\item{terms}{The term in linear algebra}
\item{Definition}{The definition for the term in linear algebra}
}
}
\source{
\url{https://www.math.ucdavis.edu/~daddel/MATH22AL/Resources/Glossary_Linear_Algebra_UMBC.html}
}
\usage{
glossary
}
\description{
Glossory of linear algebra terms
}
\keyword{datasets}
