% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{i_matrix}
\alias{i_matrix}
\title{i_matrix}
\usage{
i_matrix(size)
}
\arguments{
\item{size}{The number of rows/columns for the identity matrix}
}
\value{
The identity matrix of the given size
}
\description{
creates an identity matrix
}
\examples{
\dontrun{i_matrix(5)}

}
