% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{inverse_matrix}
\alias{inverse_matrix}
\title{inverse_matrix}
\usage{
inverse_matrix(matrix)
}
\arguments{
\item{matrix}{The matrix you want the inverse of}
}
\value{
The inverse of the given matrix
}
\description{
creates the inverse of a matrix
}
\examples{
\dontrun{inverse_matrix(matrix(1:36, ncol = 6))}

}
