% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{is_invertible}
\alias{is_invertible}
\title{is_invertible}
\usage{
is_invertible(matrix)
}
\arguments{
\item{matrix}{The matrix you want to know if it is invertible}
}
\value{
TRUE if it is invertible, FALSE if not invertible
}
\description{
tells whether a given matrix is invertible
}
\examples{
\dontrun{is_invertible(matrix(1:25, ncol = 5))}

}
