% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixReduceProcess.R
\name{isConsistent}
\alias{isConsistent}
\title{TestConsistent}
\usage{
isConsistent(Matrix, attachMatrix)
}
\arguments{
\item{Matrix}{any matrix}

\item{attachMatrix}{matrix or vector (only support vector for current version)}
}
\value{
consistent: TRUE; inconsisitent: FALSE
}
\description{
Test if a linear system is consistent or inconsistent
}
