% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-k_check.R
\name{k_check}
\alias{k_check}
\title{Sanity Checks for the k Dimension of a k by k Matrix}
\usage{
k_check(k)
}
\arguments{
\item{k}{Positive integer.
Dimension of the \code{k} by \code{k} matrix.}
}
\description{
Sanity Checks for the k Dimension of a k by k Matrix
}
\seealso{
Other Symmetric Functions: 
\code{\link{.dcap}()},
\code{\link{.mcap_cor}()},
\code{\link{.mcap_diag}()},
\code{\link{.mcap_sym}()},
\code{\link{.pinv_of_dcap}()},
\code{\link{.sym_of_vechs}()},
\code{\link{.sym_of_vech}()},
\code{\link{dcap}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{pinv_of_dcap}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{check}
\keyword{linearAlgebra}
