% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-package.R
\docType{package}
\name{linearAlgebra}
\alias{linearAlgebra}
\title{linearAlgebra: Linear Algebra}
\description{
A collection of functions related to linear algebra.
}
\section{Dot Functions}{

Functions with an initial \code{.} are functions
that perform a task without
doing sanity checks on the input arguments.
}

\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{linearAlgebra}
\keyword{package}
