% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-pinv_of_dcap.R
\name{pinv_of_dcap}
\alias{pinv_of_dcap}
\title{The Moore-Penrose Inverse of the Duplication Matrix}
\usage{
pinv_of_dcap(k)
}
\arguments{
\item{k}{Positive integer.
Dimension of the \code{k} by \code{k} matrix.}
}
\description{
The Moore-Penrose inverse of the duplication matrix.
}
\details{
The Moore-Penrose inverse of the duplication matrix
\eqn{\mathbf{D}_{k}}
is the
\eqn{
    \frac{1}{2} k \left( k + 1 \right) \times k^2
}
matrix given by

\deqn{
    \mathbf{D}_{k}^{+}
    =
    \left(
    \mathbf{D}_{k}^{\prime}
    \mathbf{D}_{k}
    \right)^{-1}
    \mathbf{D}_{k}^{\prime}
}

where

\deqn{
   \mathbf{D}_{k}^{+}
   \mathrm{vec}
   \left(
   A
   \right)
   =
   \mathrm{vech}
   \left(
   \mathbf{A}
   \right)
   \quad
   \left(
   \mathbf{A} = \mathbf{A}^{\prime}
   \right)
}

\eqn{
    \mathrm{vec} \left( \cdot \right)
}
is the vectorization of a matrix,
and
\eqn{
    \mathrm{vech} \left( \cdot \right)
}
is the half-vectorization of a matrix.
}
\examples{
pinv_of_dcap(3)
}
\seealso{
Other Symmetric Functions: 
\code{\link{dcap}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
