% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{pivot_check}
\alias{pivot_check}
\title{Check number of pivots of the matrix}
\usage{
pivot_check(matrix)
}
\arguments{
\item{matrix}{a matrix of your choice}
}
\value{
a sentence about the number of pivots of the given matrix
}
\description{
Given a matrix object, this function calculates its rref form and counted the
number of pivots, this returns a sentence of information.
}
