% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{random_matrix_generator}
\alias{random_matrix_generator}
\title{Generate a random matrix for practice}
\usage{
random_matrix_generator(ncols, nrows, min, max)
}
\arguments{
\item{ncols}{the number of columns of the matrix}

\item{nrows}{number of rows of the matrix}

\item{min}{minimum of the random number included in the matrix}

\item{max}{maximum of the random number included in the matrix}
}
\value{
x a random matrix generated
}
\description{
Given ncols, nrows, min, max, this method generates a random matrix
}
\examples{
\dontrun{
random_matrix_generator(100,100,0,100)
}
}
