% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixReduceProcess.R
\name{ReduceAugmentedMatrix}
\alias{ReduceAugmentedMatrix}
\title{Reduce Augmented Matrix}
\usage{
ReduceAugmentedMatrix(coefMatrix, attachVector, FRAC = TRUE, PRINT = FALSE)
}
\arguments{
\item{coefMatrix}{ceofficient matrix}

\item{attachVector}{An additional column which attach to the coefficient matrix => augmented matrix}

\item{FRAC}{Flag of showing result by fraction (Default is TRUE)}

\item{PRINT}{Flag of printing process detail (Default is FALSE)}
}
\value{
Return a List with StrictlyTriangularForm; AttachVector; Answer
}
\description{
Reduce Augmented Matrix to Strictly Triangular Form
Using Row Operations I, III
(only works for square matrix)
(will have the same result as solve(coefMatrix, attachVector))
}
