% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{row_replace}
\alias{row_replace}
\title{row_replace}
\usage{
row_replace(matrix, row1, row2)
}
\arguments{
\item{matrix}{The matrix where you would like a row replaced}

\item{row1}{The row you use as a reference to replace the other row}

\item{row2}{The row to be replaced}
}
\value{
The resulting matrix with the replaced row
}
\description{
replaces a row by multiplying by another row by a constant and adding it to the row you want to replace.
}
\examples{
\dontrun{row_replace(matrix(1:20, nrow = 4), 2, 4)}

}
