% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{row_swap}
\alias{row_swap}
\title{row_swap}
\usage{
row_swap(matrix, row1, row2)
}
\arguments{
\item{matrix}{The matrix you would like the rows to be swapped in}

\item{row1}{The first row you would like swapped}

\item{row2}{The second row you would like swapped}
}
\value{
The matrix that results after the rows are swapped
}
\description{
swaps two rows in a matrix
}
\examples{
\dontrun{row_swap(matrix(1:9, nrow = 3), 1, 3)}

}
