% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{rows_of_zero_detector}
\alias{rows_of_zero_detector}
\title{Given a matrix, this detect rows of zero of the matrix's rref form}
\usage{
rows_of_zero_detector(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
Given a matrix, this detect rows of zero of the matrix's rref form
}
