% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{rref}
\alias{rref}
\title{rref}
\usage{
rref(matrix)
}
\arguments{
\item{matrix}{The matrix you would like the rref of}
}
\value{
The row reduced echelon form of the given matrix
}
\description{
algorithm for getting the row reduced echelon form of a matrix
}
\examples{
\dontrun{rref(rbind(c(1, 3, 4, 6, 7), c(0, 0, 4, 5, 3), c(10, 4, 5, 8, 9)))}

}
