% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-sym_of_vech.R
\name{sym_of_vech}
\alias{sym_of_vech}
\title{Symmetric matrix A from vech(A)}
\usage{
sym_of_vech(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
A vector.
}
\description{
Symmetric matrix from its half-vectorization.
}
\details{
Generates an
\eqn{k \times k}
symmetric matrix
from a
\eqn{\frac{1}{2}k(k + 1)}
vector.
}
\examples{
A <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  ncol = 3
)
vechA <- c(1.0, 0.5, 0.4, 1.0, 0.6, 1.0)

sym_of_vech(vechA)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)#Half-vectorization}{Wikipedia: Half-vectorization}
}
\seealso{
Other Symmetric Functions: 
\code{\link{dcap}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{pinv_of_dcap}()},
\code{\link{sym_of_vechs}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
