% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-sym_of_vechs.R
\name{sym_of_vechs}
\alias{sym_of_vechs}
\title{Symmetric matrix A from vechs(A)}
\usage{
sym_of_vechs(x, diags)
}
\arguments{
\item{x}{Vector of length \code{0.5 * k(k + 1) - k}.
Strict half-vectorization of a \code{k} by \code{k} matrix.
\eqn{\mathrm{vechs} \left( \mathbf{A}_{k \times k} \right)}.}

\item{diags}{Vector.
Diagonal elements.}
}
\value{
A matrix.
}
\description{
Symmetric matrix from its strict half-vectorization.
}
\details{
Generates an
\eqn{k \times k}
symmetric matrix
from a
\eqn{\frac{1}{2}(k(k + 1)) - k}
vector.
The
\eqn{k \times 1}
vector of diagonal values should be supplied.
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)}{Wikipedia: Vectorization}

Magnus, J. R., & Neudecker, H. (2019).
Matrix  differential  calculus with  applications  in  statistics  and  econometrics.
Wiley.
\url{https://doi.org/10.1002/9781119541219}
}
\seealso{
Other Symmetric Functions: 
\code{\link{dcap}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{pinv_of_dcap}()},
\code{\link{sym_of_vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
