% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-test_positive_definite.R
\name{test_positive_definite}
\alias{test_positive_definite}
\title{Test for a Positive Definite Matrix}
\usage{
test_positive_definite(x, tol = 1e-08)
}
\arguments{
\item{x}{Numeric matrix.}

\item{tol}{Numeric. Tolerance.}
}
\value{
Logical.
}
\description{
Returns \code{TRUE} if input
is a positive definite matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} symmetric matrix \eqn{\mathbf{A}}
is positive definite
if all of its eigenvalues are positive.

\code{A} is considered to be a positive definite matrix
if \strong{NONE} of its eigenvalues are less than or equal to
a tolerance value.
}
\examples{
# TRUE
test_positive_definite(diag(2))

# FALSE
test_positive_definite(
  matrix(
    data = c(1, 2, 3, 2, 4, 5, 3, 5, 6),
    ncol = 3
  )
)
}
\references{
\href{https://en.wikipedia.org/wiki/Definite_matrix}{Wikipedia: Definite matrix}
}
\seealso{
Other Linear Algebra Functions: 
\code{\link{dcap}()},
\code{\link{diag_of_vech}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Linear Algebra Functions}
\keyword{linearAlgebra}
\keyword{test}
