% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Algebra Functions.R
\name{transpose}
\alias{transpose}
\title{transpose}
\usage{
transpose(matrix)
}
\arguments{
\item{matrix}{the matrix you would like transposed}
}
\value{
the transpose of the matrix you inputed
}
\description{
transposes a given matrix
}
\examples{
\dontrun{transpose(matrix(1:6, nrow = 2))}

}
