% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vec.R
\name{vec}
\alias{vec}
\title{Vectorize}
\usage{
vec(x)
}
\arguments{
\item{x}{Matrix.}
}
\value{
A vector.
}
\description{
Vectorize a matrix.
}
\details{
Generates a vector of length \eqn{mn}
from the elements of an \eqn{m \times n} matrix.
}
\examples{
x <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  ncol = 3
)

vec(x)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)}{Wikipedia: Vectorization}
}
\seealso{
Other Linear Algebra: 
\code{\link{dcap}()},
\code{\link{diag_of_vech}()},
\code{\link{kcap}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()},
\code{\link{test_positive_definite}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Linear Algebra}
\keyword{linearAlgebra}
\keyword{vectorization}
