% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vech.R
\name{vech}
\alias{vech}
\title{Half-Vectorize}
\usage{
vech(x, names = FALSE, sep = ".")
}
\arguments{
\item{x}{Matrix.}

\item{names}{Logical.
Add names.}

\item{sep}{Character string.
Separator for variable names.}
}
\description{
Apply a half-vectorization,
that is,
eliminate the upper diagonal elements of a
\eqn{k \times k}
matrix.
}
\details{
The half-vectorization of a
\eqn{k \times k}
matrix
\eqn{\mathbf{A}},
given by
\eqn{
    \mathrm{vech} \left( \mathbf{A} \right)
},
is the
\eqn{
    \frac{1}{2}
    k
    \left(
    k + 1
    \right)
    \times
    1
}
vector obtained from the vectorization of
\eqn{\mathbf{A}},
given by
\eqn{
    \mathrm{vec} \left( \mathbf{A} \right)
},
where that all upper diagonal elements of
\eqn{\mathbf{A}}
are eliminated.
}
\examples{
A <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  nrow = 3
)

vech(A)
}
\seealso{
Other Vectorization Functions: 
\code{\link{diag_of_vech}()},
\code{\link{vec_mean}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{linearAlgebra}
\keyword{vectorization}
