% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vech_check.R
\name{vech_check}
\alias{vech_check}
\title{Sanity Checks for the Half-Vectorization}
\usage{
vech_check(x, return_k = FALSE)
}
\arguments{
\item{x}{Vector of length \code{0.5 * k(k + 1)}.
Half-vectorization of a \code{k} by \code{k} matrix.
\eqn{\mathrm{vech} \left( \mathbf{A}_{k \times k} \right)}.}

\item{return_k}{Logical.
Return valid \code{k}.}
}
\description{
Sanity Checks for the Half-Vectorization
}
\seealso{
Other Vectorization Functions: 
\code{\link{.diag_of_vech}()},
\code{\link{.vechnames}()},
\code{\link{.vechsnames}()},
\code{\link{.vechs}()},
\code{\link{.vech}()},
\code{\link{.vec}()},
\code{\link{diag_of_vech}()},
\code{\link{vechnames}()},
\code{\link{vechs_check}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{check}
\keyword{linearAlgebra}
