% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vechs_check.R
\name{vechs_check}
\alias{vechs_check}
\title{Sanity Checks for the Strict Half-Vectorization}
\usage{
vechs_check(x, diags = NULL, return_k = FALSE)
}
\arguments{
\item{x}{Vector of length \code{0.5 * k(k + 1) - k}.
Strict half-vectorization of a \code{k} by \code{k} matrix.
\eqn{\mathrm{vechs} \left( \mathbf{A}_{k \times k} \right)}.}

\item{diags}{Vector.
Diagonal elements.}

\item{return_k}{Logical.
Return valid \code{k}.}
}
\description{
Sanity Checks for the Strict Half-Vectorization
}
\seealso{
Other Vectorization Functions: 
\code{\link{.diag_of_vech}()},
\code{\link{.vechnames}()},
\code{\link{.vechsnames}()},
\code{\link{.vechs}()},
\code{\link{.vech}()},
\code{\link{.vec}()},
\code{\link{diag_of_vech}()},
\code{\link{vech_check}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{check}
\keyword{linearAlgebra}
