% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vechsnames.R
\name{vechsnames}
\alias{vechsnames}
\title{Vector Names for Strict Half-Vectorization}
\usage{
vechsnames(x, sep = ".")
}
\arguments{
\item{x}{Character vector of names.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A character vector.
}
\description{
Vector Names for Strict Half-Vectorization
}
\examples{
x <- diag(2)
colnames(x) <- rownames(x) <- c("x1", "x2")
vechsnames(colnames(x))

x <- diag(3)
colnames(x) <- rownames(x) <- c("x1", "x2", "x3")
vechsnames(colnames(x))
}
\seealso{
Other Vectorization Functions: 
\code{\link{.diag_of_vech}()},
\code{\link{.vechs}()},
\code{\link{.vech}()},
\code{\link{.vec}()},
\code{\link{diag_of_vech}()},
\code{\link{vechnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{linearAlgebra}
\keyword{vectorization}
