% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-z.R
\name{z}
\alias{z}
\title{Standarize}
\usage{
z(x)
}
\arguments{
\item{x}{Numeric matrix or vector.
Data.}
}
\value{
A matrix.
}
\description{
Calculates standardized scores.
}
\details{
The standardized score for the
\eqn{j^{\mathrm{th}}}
column of a matrix
\eqn{\mathbf{X}}
is given by

\deqn{\frac{\mathbf{x}_{j} - \bar{x}_{j}}{s_{j}}}

where
\eqn{\mathbf{x}_{j}}
is the
\eqn{j^{\mathrm{th}}}
column of a matrix
\eqn{\mathbf{X}},
\eqn{\bar{x}_{j}}
is the mean of
\eqn{\mathbf{x}_{j}},
and
\eqn{s_{j}}
is the standard deviation of
\eqn{\mathbf{x}_{j}}.
}
\examples{
x <- matrix(
  rnorm(n = 5 * 5, mean = 100, sd = 15),
  ncol = 5
)

z(x)
}
\seealso{
Other Scaling Functions: 
\code{\link{deltacapsq}()},
\code{\link{d}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Scaling Functions}
\keyword{linearAlgebra}
\keyword{scaling}
