% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-dcap.R
\name{dcap}
\alias{dcap}
\title{The Duplication Matrix}
\usage{
dcap(x)
}
\arguments{
\item{x}{Integer.
Dimension of the symmetric matrix.}
}
\value{
A matrix.
}
\description{
Creates a duplication matrix.
}
\details{
The duplication matrix
\eqn{\mathbf{D}_{k}}
is the
\eqn{
    k^2 \times \frac{1}{2} k \left( k + 1 \right)
}
matrix for a given
\eqn{k \times k}
symmetric matrix
\eqn{\mathbf{A}}
where

\deqn{
   \mathbf{D}_{k} \mathrm{vech} \left( \mathbf{A} \right)
   =
   \mathrm{vec} \left( \mathbf{A} \right)
   \quad
   \left(
   \mathbf{A}
   =
   \mathbf{A}^{\prime}
   \right)
}

\eqn{
    \mathrm{vec} \left( \cdot \right)
}
is the vectorization of a matrix,
and
\eqn{
    \mathrm{vech} \left( \cdot \right)
}
is the half-vectorization of a matrix.
}
\examples{
dcap(3)
}
\references{
\href{https://en.wikipedia.org/wiki/Duplication_and_elimination_matrices#Duplication_matrix}{Wikipedia: Duplication matrix}

Magnus, J. R., & Neudecker, H. (1980).
The elimination matrix: Some lemmas and applications.
SIAM  Journal  on  Algebraic  Discrete  Methods,
1(4),
422--449.
\url{https://doi.org/10.1137/0601049}

Magnus, J. R., & Neudecker, H. (2019).
Matrix  differential  calculus with  applications  in  statistics  and  econometrics.
Wiley.
\url{https://doi.org/10.1002/9781119541219}
}
\seealso{
Other Symmetric Functions: 
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{pinv_of_dcap}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
