% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{linear_dependent_detector}
\alias{linear_dependent_detector}
\title{Given a matrix, this returns if the matrix is linearly dependent or not}
\usage{
linear_dependent_detector(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
Given a matrix, this returns if the matrix is linearly dependent or not
}
\examples{
c<-matrix(c(1, 2, 3,
4, 5, 6,
7, 8, 9),  # Note: this row repeats the first row
nrow=4, ncol=3, byrow=TRUE)
linear_dependent_detector(c)
}
