% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{linear_dependent_detector}
\alias{linear_dependent_detector}
\title{Given a matrix, this returns if the matrix is linearly dependent or not}
\usage{
linear_dependent_detector(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
Given a matrix, this returns if the matrix is linearly dependent or not
}
