% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-package.R
\docType{package}
\name{linearAlgebra}
\alias{linearAlgebra}
\title{linearAlgebra: Linear Algebra}
\description{
A collection of functions related to linear algebra.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{linearAlgebra}
\keyword{package}
