% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-pinv_of_dcap.R
\name{pinv_of_dcap}
\alias{pinv_of_dcap}
\title{The Moore-Penrose Inverse of the Duplication Matrix}
\usage{
pinv_of_dcap(k)
}
\arguments{
\item{k}{Positive integer.
Dimension of the \code{k} by \code{k} matrix.}
}
\value{
A matrix.
}
\description{
The Moore-Penrose inverse of the duplication matrix.
}
\details{
The Moore-Penrose inverse of the duplication matrix
\eqn{\mathbf{D}_{k}}
is the
\eqn{
    \frac{1}{2} k \left( k + 1 \right) \times k^2
}
matrix given by

\deqn{
    \mathbf{D}_{k}^{+}
    =
    \left(
    \mathbf{D}_{k}^{\prime}
    \mathbf{D}_{k}
    \right)^{-1}
    \mathbf{D}_{k}^{\prime}
}

where

\deqn{
   \mathbf{D}_{k}^{+}
   \mathrm{vec}
   \left(
   A
   \right)
   =
   \mathrm{vech}
   \left(
   \mathbf{A}
   \right)
   \quad
   \left(
   \mathbf{A} = \mathbf{A}^{\prime}
   \right)
}

\eqn{
    \mathrm{vec} \left( \cdot \right)
}
is the vectorization of a matrix,
and
\eqn{
    \mathrm{vech} \left( \cdot \right)
}
is the half-vectorization of a matrix.
}
\examples{
pinv_of_dcap(3)
}
\references{
\href{https://en.wikipedia.org/wiki/Duplication_and_elimination_matrices#Duplication_matrix}{Wikipedia: Duplication matrix}

\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)}{Wikipedia: Vectorization}

Magnus, J. R., & Neudecker, H. (1980).
The elimination matrix: Some lemmas and applications.
SIAM  Journal  on  Algebraic  Discrete  Methods,
1(4),
422--449.
\url{https://doi.org/10.1137/0601049}

Magnus, J. R., & Neudecker, H. (2019).
Matrix  differential  calculus with  applications  in  statistics  and  econometrics.
Wiley.
\url{https://doi.org/10.1002/9781119541219}
}
\seealso{
Other Symmetric Functions: 
\code{\link{.dcap}()},
\code{\link{.mcap_cor}()},
\code{\link{.mcap_diag}()},
\code{\link{.mcap_sym}()},
\code{\link{.pinv_of_dcap}()},
\code{\link{.sym_of_vechs}()},
\code{\link{.sym_of_vech}()},
\code{\link{dcap}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()}
}
\concept{Symmetric Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
