% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-sym_of_vechs.R
\name{sym_of_vechs}
\alias{sym_of_vechs}
\title{Symmetric matrix A from vechs(A)}
\usage{
sym_of_vechs(x, diags)
}
\arguments{
\item{x}{Vector.}

\item{diags}{Vector.
Diagonal elements.}
}
\value{
A vector.
}
\description{
Symmetric matrix from its strict half-vectorization.
}
\details{
Generates an
\eqn{k \times k}
symmetric matrix
from a
\eqn{\frac{1}{2}(k(k + 1)) - k}
vector.
The
\eqn{k \times 1}
vector of diagonal values should be supplied.
}
\examples{
A <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  ncol = 3
)
vechsA <- c(0.5, 0.4, 0.6)

sym_of_vechs(vechsA, diags = 1)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)#Half-vectorization}{Wikipedia: Half-vectorization}
}
\seealso{
Other Symmetric Functions: 
\code{\link{.dcap}()},
\code{\link{.mcap_cor}()},
\code{\link{.mcap_diag}()},
\code{\link{.mcap_sym}()},
\code{\link{.pinv_of_dcap}()},
\code{\link{.sym_of_vechs}()},
\code{\link{.sym_of_vech}()},
\code{\link{dcap}()},
\code{\link{mcap_cor}()},
\code{\link{mcap_diag}()},
\code{\link{mcap_sym}()},
\code{\link{pinv_of_dcap}()},
\code{\link{sym_of_vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Symmetric Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
