% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vec_mean.R
\name{vec_mean}
\alias{vec_mean}
\title{Vector Mean}
\usage{
vec_mean(x)
}
\arguments{
\item{x}{Numeric vector or \code{n} by \code{1} matrix.}
}
\value{
A vector of length 1.
}
\description{
Calculates the mean of a vector.
If the input is a matrix,
the function vectorizes the matrix
before calculating the mean.
}
\details{
The vector mean is given by

\deqn{
  \frac{1}{n}
  \left(
  \mathbf{1}_{n}^{\prime}
  \mathbf{x}
  \right)
}
}
\examples{
x <- rnorm(100)
vec_mean(x)
vec_mean(matrix(x, ncol = 1))
}
\seealso{
Other Vectorization Functions: 
\code{\link{diag_of_vech}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()},
\code{\link{vec}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Vectorization Functions}
\keyword{linearAlgebra}
\keyword{operation}
